hotfix(sqlmesh): rounding issues again (#5426)

+++ warehouse/oso_sqlmesh/models/intermediate/domain-specific/superchain/grant-measurement/int_optimism_grants_rolling_defi_metrics_by_chain.sql
@@ -34,7 +34,7 @@ grouped_metrics AS (
     sample_date,
     chain,
     metric,
-    SUM(amount)::DOUBLE AS amount
+    SUM(amount) AS amount
   FROM base
   GROUP BY 1,2,3
 ),
